OTA Agent Trap for Stack in External Memory
A simple trap to detect and report if the OTA Agent Task's stack is allocated from external memory.
When this allocation occurs the system becomes unstable.  Typical symptoms include TASK_WDT resets
and double exceptions.  The work around has been to reduce the size of other task stacks so that
the OTA Agent's stack is allocated from internal memory.

The code is commented out, for release, but retained for debugging future issues.